+++ src/components/calcite-button/calcite-button.stories.js
@@ -2,8 +2,14 @@ import { storiesOf } from "@storybook/html";
 import { withKnobs, text, boolean, select } from "@storybook/addon-knobs";
 import { images24 } from "@esri/calcite-ui-icons/js/images24";
 import { darkBackground, parseReadme } from "../../../.storybook/helpers";
+import * as icons from "../../../node_modules/@esri/calcite-ui-icons";
 import readme from "./readme.md";
 const notes = parseReadme(readme);
+
+// we can get all unique icon names from all size 16 non-filled icons.
+const iconNames = Object.keys(icons)
+  .filter(iconName => iconName.endsWith("16"))
+  .map(iconName => iconName.replace("16", ""));
 
 storiesOf("Button", module)
   .addDecorator(withKnobs)
@@ -45,7 +51,7 @@ storiesOf("Button", module)
       href="${text("href", "")}"
       loading="${boolean("loading", false)}"
       disabled="${boolean("disabled", false)}"
-      icon="${text("icon", images24)}"
+      icon="${select("icon", iconNames, iconNames[0])}"
       icon-position="${select("icon-position", ["start", "end"], "start")}">
       ${text("text", "button text here")}
     </calcite-button>
@@ -58,7 +64,7 @@ storiesOf("Button", module)
     <div style="width: 480px; max-width: 100%; background-color: #fff">
       <calcite-button
         width="${select("width", ["auto", "half", "full"], "auto")}"
-        icon="${text("icon", images24)}"
+        icon="${select("icon", iconNames, iconNames[0])}"
         icon-position="${select("icon-position", ["start", "end"], "start")}">
         ${text("text", "button text here")}
       </calcite-button>
@@ -88,7 +94,7 @@ storiesOf("Button", module)
       "solid"
     )}"
     color="${select("color-2", ["blue", "red", "dark", "light"], "blue")}"
-    icon="${text("icon", images24)}">
+    icon="${select("icon", iconNames, iconNames[0])}">
     ${text("text-2", "Some long string")}
     </calcite-button>
   </div>
@@ -122,7 +128,7 @@ storiesOf("Button", module)
   round="${boolean("round", true)}"
   floating="${boolean("floating", true)}"
   width="${select("width", ["auto", "half", "full"], "auto")}"
-  icon="${text("icon", images24)}"
+  icon="${select("icon", iconNames, iconNames[0])}"
   appearance="${select("appearance", ["solid", "outline"], "solid")}"
   color="${select("color", ["blue", "red", "dark", "light"], "blue")}"
   scale="${select("scale", ["xs", "s", "m", "l", "xl"], "m")}"></calcite-button>
@@ -159,7 +165,7 @@ storiesOf("Button", module)
     round="${boolean("round", true)}"
     floating="${boolean("floating", true)}"
     width="${select("width", ["auto", "half", "full"], "auto")}"
-    icon="${text("icon", images24)}"
+    icon="${select("icon", iconNames, iconNames[0])}"
     appearance="${select("appearance", ["solid", "outline"], "solid")}"
     color="${select("color", ["blue", "red", "dark", "light"], "blue")}"
     scale="${select("scale", ["xs", "s", "m", "l", "xl"], "m")}">
@@ -187,7 +193,7 @@ storiesOf("Button", module)
     href="${text("href", "")}"
     loading="${boolean("loading", false)}"
     disabled="${boolean("disabled", false)}"
-    icon="${text("icon", images24)}">
+    icon="${select("icon", iconNames, iconNames[0])}">
     ${text("text", "button text here")}
   </calcite-button>
   `,
@@ -221,11 +227,10 @@ storiesOf("Button", module)
     round="${boolean("round", true)}"
     floating="${boolean("floating", true)}"
     width="${select("width", ["auto", "half", "full"], "auto")}"
-    icon="${text("icon", images24)}"
+    icon="${select("icon", iconNames, iconNames[0])}"
     appearance="${select("appearance", ["solid", "outline"], "solid")}"
     color="${select("color", ["blue", "red", "dark", "light"], "blue")}"
-    scale="${select("scale", ["xs", "s", "m", "l", "xl"], "m")}">
-    </calcite-button>
+    scale="${select("scale", ["xs", "s", "m", "l", "xl"], "m")}"></calcite-button>
   </div>
 </div>
 <calcite-tooltip theme="dark" reference-element="calcite-fab-tooltip">Add new</calcite-tooltip>
@@ -259,7 +264,7 @@ storiesOf("Button", module)
   round="${boolean("round", true)}"
   floating="${boolean("floating", true)}"
   width="${select("width", ["auto", "half", "full"], "auto")}"
-  icon="${text("icon", images24)}"
+  icon="${select("icon", iconNames, iconNames[0])}"
   appearance="${select("appearance", ["solid", "outline"], "solid")}"
   color="${select("color", ["blue", "red", "dark", "light"], "blue")}"
   scale="${select("scale", ["xs", "s", "m", "l", "xl"], "m")}">