+++ src/components/calcite-dropdown-item/calcite-dropdown-item.tsx
@@ -18,11 +18,7 @@ import {
   END,
   SPACE
 } from "../../utils/keys";
-import {
-  getElementDir,
-  getElementTheme,
-  getElementProp
-} from "../../utils/dom";
+import { getElementDir, getElementProp } from "../../utils/dom";
 import { guid } from "../../utils/guid";
 
 @Component({
@@ -47,12 +43,14 @@ export class CalciteDropdownItem {
 
   @Prop({ reflect: true, mutable: true }) active: boolean = false;
 
-  /** pass an optional href to render an anchor around the link items */
-  @Prop() href?: string;
+  /** optionally pass an icon to display at the start of an item - accepts calcite ui icon names  */
+  @Prop({ reflect: true }) iconStart?: string;
 
-  /** pass an optional title for rendered href */
-  @Prop() linkTitle?: string;
+  /** optionally pass an icon to display at the end of an item - accepts calcite ui icon names  */
+  @Prop({ reflect: true }) iconEnd?: string;
 
+  /** optionally pass a href - used to determine if the component should render as anchor */
+  @Prop({ reflect: true }) href?: string;
   //--------------------------------------------------------------------------
   //
   //  Events
@@ -79,39 +77,60 @@ export class CalciteDropdownItem {
   }
 
   render() {
+    const attributes = this.getAttributes();
     const dir = getElementDir(this.el);
-    const theme = getElementTheme(this.el);
     const scale = getElementProp(this.el, "scale", "m");
-    if (!this.href) {
-      return (
-        <Host
-          theme={theme}
-          dir={dir}
-          scale={scale}
-          tabindex="0"
-          role="menuitem"
-          aria-selected={this.active.toString()}
-        >
-          <slot />
-        </Host>
+    const iconScale = scale === "s" || scale === "m" ? "s" : "m";
+    const iconStartEl = (
+      <calcite-icon
+        class="dropdown-item-icon-start"
+        icon={this.iconStart}
+        scale={iconScale}
+      ></calcite-icon>
+    );
+    const iconEndEl = (
+      <calcite-icon
+        class="dropdown-item-icon-end"
+        icon={this.iconEnd}
+        scale={iconScale}
+      ></calcite-icon>
+    );
+
+    const slottedContent =
+      this.iconStart && this.iconEnd ? (
+        [iconStartEl, <slot />, iconEndEl]
+      ) : this.iconStart ? (
+        [iconStartEl, <slot />]
+      ) : this.iconEnd ? (
+        [<slot />, iconEndEl]
+      ) : (
+        <slot />
       );
-    } else {
-      return (
-        <Host
-          theme={theme}
-          dir={dir}
-          scale={scale}
-          tabindex="0"
-          role="menuitem"
-          aria-selected={this.active.toString()}
-          isLink
-        >
-          <a href={this.href} title={this.linkTitle}>
-            <slot />
-          </a>
-        </Host>
-      );
-    }
+
+    const contentEl = !this.href ? (
+      slottedContent
+    ) : (
+      <a {...attributes}>{slottedContent}</a>
+    );
+    return (
+      <Host
+        dir={dir}
+        tabindex="0"
+        role="menuitem"
+        selection-mode={this.selectionMode}
+        aria-selected={this.active.toString()}
+        isLink={this.href}
+      >
+        {this.selectionMode === "multi" ? (
+          <calcite-icon
+            class="dropdown-item-check-icon"
+            scale="s"
+            icon="check"
+          />
+        ) : null}
+        {contentEl}
+      </Host>
+    );
   }
 
   //--------------------------------------------------------------------------
@@ -215,6 +234,23 @@ export class CalciteDropdownItem {
       requestedDropdownGroup: this.currentDropdownGroup
     });
     this.closeCalciteDropdown.emit();
+  }
+
+  private getAttributes() {
+    // spread attributes from the component to rendered child, filtering out props
+    let props = [
+      "icon-start",
+      "icon-end",
+      "active",
+      "hasText",
+      "isLink",
+      "dir",
+      "id",
+      "theme"
+    ];
+    return Array.from(this.el.attributes)
+      .filter(a => a && !props.includes(a.name))
+      .reduce((acc, { name, value }) => ({ ...acc, [name]: value }), {});
   }
 
   private getItemPosition() {