+++ src/components/calcite-dropdown/calcite-dropdown.e2e.ts
@@ -44,15 +44,16 @@ describe("calcite-dropdown", () => {
     const element = await page.find("calcite-dropdown");
     const group1 = await element.find("calcite-dropdown-group[id='group-1']");
     expect(element).toEqualAttribute("scale", "m");
+    expect(element).toEqualAttribute("width", "m");
     expect(element).toEqualAttribute("theme", "light");
-    expect(element).toEqualAttribute("alignment", "left");
+    expect(element).toEqualAttribute("alignment", "start");
     expect(group1).toEqualAttribute("selection-mode", "single");
   });
 
   it("renders default props when invalid props are provided", async () => {
     const page = await newE2EPage();
     await page.setContent(`
-    <calcite-dropdown alignment="zip" scale="zop" theme="zut">
+    <calcite-dropdown alignment="zip" scale="zop" theme="zut" width="zat">
     <calcite-button slot="dropdown-trigger">Open dropdown</calcite-button>
     <calcite-dropdown-group id="group-1" selection-mode="zap">
     <calcite-dropdown-item id="item-1">
@@ -70,15 +71,16 @@ describe("calcite-dropdown", () => {
     const element = await page.find("calcite-dropdown");
     const group1 = await element.find("calcite-dropdown-group[id='group-1']");
     expect(element).toEqualAttribute("scale", "m");
+    expect(element).toEqualAttribute("width", "m");
     expect(element).toEqualAttribute("theme", "light");
-    expect(element).toEqualAttribute("alignment", "left");
+    expect(element).toEqualAttribute("alignment", "start");
     expect(group1).toEqualAttribute("selection-mode", "single");
   });
 
   it("renders requested props when valid props are provided", async () => {
     const page = await newE2EPage();
     await page.setContent(`
-    <calcite-dropdown alignment="right" scale="l" theme="dark">
+    <calcite-dropdown alignment="end" scale="l" width="l" theme="dark">
     <calcite-button slot="dropdown-trigger">Open dropdown</calcite-button>
     <calcite-dropdown-group id="group-1" selection-mode="multi">
     <calcite-dropdown-item id="item-1">
@@ -96,9 +98,65 @@ describe("calcite-dropdown", () => {
     const element = await page.find("calcite-dropdown");
     const group1 = await element.find("calcite-dropdown-group[id='group-1']");
     expect(element).toEqualAttribute("scale", "l");
+    expect(element).toEqualAttribute("width", "l");
     expect(element).toEqualAttribute("theme", "dark");
-    expect(element).toEqualAttribute("alignment", "right");
+    expect(element).toEqualAttribute("alignment", "end");
     expect(group1).toEqualAttribute("selection-mode", "multi");
+  });
+
+  it("renders icons if requested and does not render icons if not requested", async () => {
+    const page = await newE2EPage();
+    await page.setContent(`
+    <calcite-dropdown>
+    <calcite-button slot="dropdown-trigger">Open dropdown</calcite-button>
+    <calcite-dropdown-group>
+    <calcite-dropdown-item icon-start="grid" id="item-1">
+    Dropdown Item Content
+    </calcite-dropdown-item>
+    <calcite-dropdown-item icon-end="grid" id="item-2" active>
+    Dropdown Item Content
+    </calcite-dropdown-item>
+    <calcite-dropdown-item icon-start="grid" icon-end="grid" id="item-3">
+    Dropdown Item Content
+    </calcite-dropdown-item>
+    <calcite-dropdown-item id="item-4">
+    Dropdown Item Content
+    </calcite-dropdown-item>
+    </calcite-dropdown-group>
+    </calcite-dropdown>`);
+
+    const item1IconStart = await page.find(
+      "calcite-dropdown-item[id='item-1'] >>> .dropdown-item-icon-start"
+    );
+    const item1IconEnd = await page.find(
+      "calcite-dropdown-item[id='item-1'] >>> .dropdown-item-icon-end"
+    );
+    const item2IconStart = await page.find(
+      "calcite-dropdown-item[id='item-2'] >>> .dropdown-item-icon-start"
+    );
+    const item2IconEnd = await page.find(
+      "calcite-dropdown-item[id='item-2'] >>> .dropdown-item-icon-end"
+    );
+    const item3IconStart = await page.find(
+      "calcite-dropdown-item[id='item-3'] >>> .dropdown-item-icon-start"
+    );
+    const item3IconEnd = await page.find(
+      "calcite-dropdown-item[id='item-3'] >>> .dropdown-item-icon-end"
+    );
+    const item4IconStart = await page.find(
+      "calcite-dropdown-item[id='item-4'] >>> .dropdown-item-icon-start"
+    );
+    const item4IconEnd = await page.find(
+      "calcite-dropdown-item[id='item-4'] >>> .dropdown-item-icon-end"
+    );
+    expect(item1IconStart).not.toBeNull();
+    expect(item1IconEnd).toBeNull();
+    expect(item2IconStart).toBeNull();
+    expect(item2IconEnd).not.toBeNull();
+    expect(item3IconStart).not.toBeNull();
+    expect(item3IconEnd).not.toBeNull();
+    expect(item4IconStart).toBeNull();
+    expect(item4IconEnd).toBeNull();
   });
 
   it("renders group title if specified and not if absent", async () => {
@@ -348,5 +406,30 @@ describe("calcite-dropdown", () => {
     expect(item7).not.toHaveAttribute("active");
     expect(item8).not.toHaveAttribute("active");
     expect(item9).not.toHaveAttribute("active");
+  });
+
+  it("renders a calcite-dropdown-item with child anchor link with passed attributes if href is present", async () => {
+    const page = await newE2EPage();
+    await page.setContent(
+      `<calcite-dropdown>
+      <calcite-button id="trigger" slot="dropdown-trigger">Open dropdown</calcite-button>
+      <calcite-dropdown-group id="group-1" selection-mode="none">
+      <calcite-dropdown-item id="item-1">
+      Dropdown Item Content
+      </calcite-dropdown-item>
+      <calcite-dropdown-item id="item-2" href="google.com" rel="noopener noreferrer" target="_blank">
+      Dropdown Item Content
+      </calcite-dropdown-item>
+      <calcite-dropdown-item id="item-3">
+      Dropdown Item Content
+      </calcite-dropdown-item>
+      </calcite-dropdown-group>
+      </calcite-dropdown>`
+    );
+    const elementAsLink = await page.find("calcite-dropdown-item[id='item-2'] >>> a");
+    expect(elementAsLink).not.toBeNull();
+    expect(elementAsLink).toEqualAttribute("href", "google.com");
+    expect(elementAsLink).toEqualAttribute("rel", "noopener noreferrer");
+    expect(elementAsLink).toEqualAttribute("target", "_blank");
   });
 });