+++ src/components/calcite-dropdown/calcite-dropdown.tsx
@@ -36,15 +36,18 @@ export class CalciteDropdown {
 
   /** specify the alignment of dropdrown, defaults to left */
   @Prop({ mutable: true, reflect: true }) alignment:
-    | "left"
-    | "right"
-    | "center" = "left";
+    | "start"
+    | "center"
+    | "end" = "start";
 
   /** specify the theme of the dropdown, defaults to light */
   @Prop({ mutable: true, reflect: true }) theme: "light" | "dark" = "light";
 
   /** specify the scale of dropdrown, defaults to m */
   @Prop({ mutable: true, reflect: true }) scale: "s" | "m" | "l" = "m";
+
+  /** specify the width of dropdrown, defaults to m */
+  @Prop({ mutable: true, reflect: true }) width: "s" | "m" | "l" = "m";
 
   /** specify whether the dropdown is opened by hover or click of the trigger element */
   @Prop({ mutable: true, reflect: true }) type: "hover" | "click" = "click";
@@ -57,14 +60,17 @@ export class CalciteDropdown {
 
   connectedCallback() {
     // validate props
-    let alignment = ["left", "right", "center"];
-    if (!alignment.includes(this.alignment)) this.alignment = "left";
+    let alignment = ["start", "center", "end"];
+    if (!alignment.includes(this.alignment)) this.alignment = "start";
 
     let theme = ["light", "dark"];
     if (!theme.includes(this.theme)) this.theme = "light";
 
     let scale = ["s", "m", "l"];
     if (!scale.includes(this.scale)) this.scale = "m";
+
+    let width = ["s", "m", "l"];
+    if (!width.includes(this.width)) this.width = "m";
 
     let type = ["hover", "click"];
     if (!type.includes(this.type)) this.type = "hover";