+++ src/components/calcite-tooltip/resources.ts
@@ -1,5 +1,4 @@
 export const CSS = {
-  container: "tooltip-container",
-  containerOpen: "tooltip-container--open",
-  contentContainer: "tooltip-content-container"
+  container: "container",
+  arrow: "arrow"
 };