+++ src/utils/popper.ts
@@ -1,4 +1,9 @@
-import Popper from "popper.js";
+import {
+  Placement,
+  Instance as Popper,
+  createPopper as setupPopper,
+  Modifier
+} from "@popperjs/core";
 import { getElementDir } from "../utils/dom";
 
 type PlacementRtl =
@@ -9,14 +14,12 @@ type PlacementRtl =
   | "trailing"
   | "trailing-start";
 
-export type CalcitePlacement = Popper.Placement | PlacementRtl;
-
-export type CalciteFlipPlacements = Popper.Position[];
+export type CalcitePlacement = Placement | PlacementRtl;
 
 export function getPlacement(
   el: HTMLElement,
   placement: CalcitePlacement
-): Popper.Placement {
+): Placement {
   const values = ["left", "right"];
 
   if (getElementDir(el) === "rtl") {
@@ -25,5 +28,57 @@ export function getPlacement(
 
   return placement
     .replace(/leading/gi, values[0])
-    .replace(/trailing/gi, values[1]) as Popper.Placement;
+    .replace(/trailing/gi, values[1]) as Placement;
 }
+
+export function createPopper({
+  referenceEl,
+  el,
+  open,
+  placement,
+  modifiers
+}: {
+  el: HTMLElement;
+  modifiers: Partial<Modifier<any>>[];
+  open: boolean;
+  placement: CalcitePlacement;
+  referenceEl: HTMLElement;
+}): Popper | null {
+  if (!referenceEl || !open) {
+    return null;
+  }
+
+  return setupPopper(referenceEl, el, {
+    placement: getPlacement(el, placement),
+    modifiers
+  });
+}
+
+export function updatePopper({
+  el,
+  modifiers,
+  placement: calcitePlacement,
+  popper
+}: {
+  el: HTMLElement;
+  modifiers: Partial<Modifier<any>>[];
+  popper: Popper;
+  placement: CalcitePlacement;
+}): void {
+  const placement = getPlacement(el, calcitePlacement);
+
+  popper.setOptions({
+    modifiers,
+    placement
+  });
+}
+
+export function hypotenuse(sideA: number, sideB: number): number {
+  return Math.sqrt(Math.pow(sideA, 2) + Math.pow(sideB, 2));
+}
+
+const visiblePointerSize = 4;
+
+export const defaultOffsetDistance = Math.ceil(
+  hypotenuse(visiblePointerSize, visiblePointerSize)
+);